fix(daemon): use resolveHomeDir in legacy lock cleanup
