+++ src/server-daemon.ts
@@ -13,6 +13,7 @@ import {
   writeServerLock,
 } from "./daemon-ops";
 import { PRIVATE_FILE_MODE } from "./file-ops";
+import { resolveHomeDir } from "./home-dir";
 import { t } from "./i18n";
 import { PROTOCOL_VERSION } from "./protocol";
 
@@ -160,7 +161,7 @@ async function waitForHealthyServerOrStaleStartupLock(
 const DEFAULT_PORT = 6767;
 
 async function cleanupLegacyLocks(homeDir?: string): Promise<void> {
-  const home = homeDir ?? "";
+  const home = homeDir ?? resolveHomeDir();
   await rm(join(home, ".acolyte", "server.lock"), { force: true });
   await rm(join(home, ".acolyte", "server.start.lock"), { force: true });
   await rm(join(daemonsDir(homeDir), `${DEFAULT_PORT}.lock`), { force: true });
